refactor(agent): uniswap llm features
