+++ packages/server/src/agent/tools/uniswap/index.ts
@@ -1,14 +1,14 @@
 import { ConfigService } from '@nestjs/config';
-import { makeRebalancePortfolioTool } from './rebalance';
-import { makePortfolioAnalyzerTool } from './portfolio-analyzer';
-import { makeLiquidityProviderTool } from './liquidity-provider';
-import { makeSuggestSwapTool } from './suggest-swap';
+import { makeRebalancePortfolioTool } from './features/rebalance';
+import { makePortfolioAnalyzerTool } from './features/portfolio-analyzer';
+import { makeLiquidityProviderTool } from './features/liquidity-provider';
+import { makeSuggestSwapTool } from './features/suggest-swap';
 
 export function makeUniswapTools(configService: ConfigService) {
   const rebalancePortfolioTool = makeRebalancePortfolioTool(configService);
   const portfolioAnalyzerTool = makePortfolioAnalyzerTool(configService);
   const liquidityProviderTool = makeLiquidityProviderTool(configService);
-  const suggestSwapTool = makeSuggestSwapTool(configService);
+  const suggestSwapTool = makeSuggestSwapTool();
 
   return {
     rebalancePortfolioTool,
