+++ packages/website/components/widget-installation-section.tsx
@@ -2,10 +2,9 @@
 
 import { Button } from '@/components/ui/button';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { Code, Copy, Check, Package, Play, BookOpen } from 'lucide-react';
+import { Code, Copy, Check, Package, BookOpen } from 'lucide-react';
 import { useState } from 'react';
 import { motion } from 'framer-motion';
-import { Separator } from './ui/separator';
 
 // Terminal command component for displaying installation commands
 function TerminalCommand({ command }: { command: string }) {
